Revert "tools/modeleditor: Have a tools tab and a compute normal button which creates smooth normals (tested with landscapes)"
This reverts commit 00f6121ebb54becb92fdf924886c0b17d5728063.

+++ src/tdme/tools/shared/files/ModelMetaDataFileExport.cpp
@@ -529,6 +529,6 @@ tdme::ext::jsonbox::Object ModelMetaDataFileExport::exportToJSON(LevelEditorEnti
 	jEntityRoot["rg"] = entity->isRenderGroups();
 	jEntityRoot["s"] = entity->getShader();
 	jEntityRoot["sds"] = entity->getDistanceShader();
-	jEntityRoot["sdsd"] = static_cast< double >(entity->getDistanceShaderDistance());
+	jEntityRoot["sdsd"] = entity->getDistanceShaderDistance();
 	return jEntityRoot;
 }
